pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vmlx"
version = "1.3.93"
description = "Local AI inference for Apple Silicon — Text, Image, Video & Audio generation on Mac"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11,<3.15"
authors = [
    {name = "Jinho Jang", email = "eric@jangq.ai"}
]
keywords = ["llm", "mlx", "apple-silicon", "vllm", "inference", "transformers"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mlx>=0.29.0",
    "mlx-lm>=0.31.2",  # 0.31.2 ships ArraysCache.lengths/advance, SequenceStateMachine,
                       # LRUPromptCache, and native Gemma 4 text MoE — vMLX audit 2026-04-07
                       # ports each. NOTE: mlx-audio 0.4.1 hard-pins 0.31.1; audio support
                       # is opt-in via [audio] extra and not in the main bundle.
    "mlx-vlm>=0.4.3",  # Gemma 4 model class added in 0.4.3
    "transformers>=4.40.0",
    "tokenizers>=0.19.0",
    "huggingface-hub>=0.23.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "tabulate>=0.9.0",
    # Video processing for VLM (headless — no GUI deps, smaller install)
    "opencv-python-headless>=4.8.0",
    # Resource monitoring
    "psutil>=5.9.0",
    # Server
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    # MCP (Model Context Protocol) support
    "mcp>=1.0.0",
    # JSON Schema validation for structured output
    "jsonschema>=4.0.0",
    # Note: mlx-audio moved to optional [audio] deps due to mlx-lm version conflict
    "mlx-embeddings>=0.0.5",
]

[project.optional-dependencies]
ui = [
    "gradio>=4.0.0",
    "pytz>=2024.1",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
vllm = [
    # vllm pins torch<=2.6.0 which conflicts with mflux (torch>=2.8.0) when both installed
    "vllm>=0.4.0; python_version < '3.13'",
]
vision = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
]
# Audio dependencies for TTS/STT (mlx-audio)
audio = [
    "mlx-audio>=0.2.9",
    "sounddevice>=0.4.0",
    "soundfile>=0.12.0",
    "scipy>=1.10.0",
    "numba>=0.57.0",
    "tiktoken>=0.5.0",
    "misaki[zh,ja]>=0.5.0",  # Chinese (zh) and Japanese (ja) support
    "spacy>=3.7.0",
    "num2words>=0.5.0",
    "loguru>=0.7.0",
    "phonemizer>=3.2.0",
    # Additional multilingual dependencies
    "ordered_set>=4.1.0",    # Required for Chinese TTS
    "cn2an>=0.5.0",          # Chinese number conversion
    "fugashi>=1.3.0",        # Japanese tokenizer
    "unidic-lite>=1.0.0",    # Japanese dictionary for fugashi
    "jieba>=0.42.0",         # Chinese word segmentation
]
# JANG conversion (inference loader is built-in, only conversion needs this)
jang = [
    "jang>=1.0.0; python_version >= '3.11'",
]
# Image generation dependencies (Flux, Z-Image-Turbo via mflux)
image = [
    "mflux>=0.16.0",
]

[project.urls]
Homepage = "https://github.com/jjang-ai/vmlx"
Documentation = "https://github.com/jjang-ai/vmlx#readme"
Repository = "https://github.com/jjang-ai/vmlx"
Downloads = "https://github.com/jjang-ai/mlxstudio/releases"
Issues = "https://github.com/jjang-ai/vmlx/issues"

[project.entry-points."vllm.platform_plugins"]
mlx = "vmlx_engine.plugin:mlx_platform_plugin"

[project.scripts]
vmlx = "vmlx_engine.cli:main"
vmlx-serve = "vmlx_engine.cli:main"
vmlx-engine = "vmlx_engine.cli:main"
vmlx-engine-chat = "vmlx_engine.gradio_app:main"
vmlx-engine-bench = "vmlx_engine.benchmark:main"
vmlx-worker = "vmlx_engine.distributed.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vmlx_engine*"]

[tool.poetry]
package-mode = false

[tool.black]
line-length = 88
target-version = ["py311", "py312", "py313"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501", "B905"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
